Implemented input output objects for each node
This will allow node functions to be implement a smaller amount of boiler plate. And allow for lists of inputs / outputs for each node. This came with a small downside that inputs and outputs now need to be manually removed from the save file to avoid a duplicate object. The data wizard class may have or eventually offer a better solution.