Edited compute section for standard and random arrays
Edited the compute portions of the standard and random arrays. Rand will now automatically convert floats to integer for set and count. Array was needing to have count set 1 higher than the actual length of the resulting array. This has now been fixed. Also updated the node gui screenshot for the readme.